Cast pk to int for cache key
I'm working in a project where the foreign key is nullable. When pandas initially imports the foreign key, the type is cast to float so the null fields can be represented as NaN. The float fields are then formatted as "4.0" and "5.0" in the cache key which cause cache misses later when the other pk's are formatted as "4" and "5" respectively.

The cast to integer here is heavy-handed but I'm not sure a better way.